[rename/#180] index.tsx를 Carousel.tsx로 변경
