+++ src/components/SearchableLayout/index.tsx
@@ -1,7 +1,7 @@
 import Magnifier from "@/assets/svgs/ic_magnifier.svg";
 import Leading from "@/assets/svgs/leading.svg";
 import Button from "@/src/components/Button/Button";
-import Carousel from "@/src/components/Carousel";
+import Carousel from "@/src/components/Carousel/Carousel";
 import items from "@/src/constants/carousel";
 import useResponsiveTextStyle from "@/src/hooks/Activity/useResponsiveTextStyle";
 import clsx from "clsx";
